feat: begin work on filters!
Woah, a lot to do here, but basically worked on familiarizing myself with firebase
and the .where() methods to query data and show it on the front end. In this commit
I just worked on the 'projectCategory' but will certainly need to feed in multiple
queries using a .reduce() function. I also added a reset button.